+++ podman-pilot/src/podman.rs
@@ -29,7 +29,8 @@ use std::env;
 use std::fs;
 use crate::config::{RuntimeSection, config};
 use crate::defaults::debug;
-use crate::error::{FlakeError, CommandError, CommandExtTrait};
+use flakes::error::FlakeError;
+use flakes::command::{CommandError, CommandExtTrait};
 use tempfile::tempfile;
 use std::io::{Write, Read};
 use std::fs::File;
